% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_march.R
\name{extract_march}
\alias{extract_march}
\title{extract_march}
\usage{
extract_march(postal_code = "20010")
}
\arguments{
\item{postal_code}{supply a single zip code location (as numeric value)}
}
\value{
tbl_df data.frame object of event locations
}
\description{
Function scrapes content from
https://event.marchforourlives.com/event/march-our-lives-events/search/ and
provides event location data for all events within fifty miles of an
organized event. All event data is as provided by protest organizers.
}
