% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/march_info_all.R
\name{march_info_all}
\alias{march_info_all}
\title{march_info_all}
\usage{
march_info_all(focal_zip = 66101)
}
\arguments{
\item{focal_zip}{this is a focal location information which all locations 20k
miles around are tracked by. Current default is somewhere in Kansas.}
}
\description{
Gather protest information on all March for Our Lives protests worldwide.
}
\examples{

all_planned_protests <- march_info_all()
all_planned_protests

}
